Move console.* monkeypatching to src/test/testLogging.ts
